backend-repo/requirements.txt
Django==4.1
djangorestframework==3.13.1
python-dotenv==0.21.0
django-redis==5.2.0
gunicorn==20.1.
django-storages==1.13.2
